Changes in the validation aspects module
- Remove Hibernate Validator as a dependency with compile scope
- Add the javax.validation api JAR as a dependency with compile scope
- Add Hibernate validator, Hibernate validator annotation processor and javax.el dependencies with runtime scope

This change allows to use the javax.validation API in the ConstraintValidationAspect class without really depending on the Hibernate's implementation. However, the basic Hibernate's stuff is included as a runtime dependency to allow consumers of this module to include it without having to add anything else. If another implementation is required, the consumer can exclude this three dependencies

Closes #24